ha_gateway/requirements.txt
aiohttp>=3.8.0
asyncio
pyyaml>=6.0
websockets>=10.0